+++ mscurrency/src/main/java/br/com/mscurrency/dto/UpdateCurrencyPriceInput.java
@@ -1,8 +1,9 @@
 package br.com.mscurrency.dto;
+
 import lombok.Data;
 
 @Data
 public class UpdateCurrencyPriceInput {
     private String currencyCode;
-    private Long priceInCentsBRL;
+    private Integer priceInCentsBRL;
 }
